+++ gptcli/src/api_helper.py
@@ -4,63 +4,50 @@ import json
 import logging
 import os
 from http import HTTPStatus
-from typing import Dict, List
+from logging import Logger
+from typing import Union
 
 import requests
+from requests import Response
 from requests.exceptions import ReadTimeout
 
-logger = logging.getLogger(__name__)
+from gptcli.src.message import Messages
+
+logger: Logger = logging.getLogger(__name__)
 
 
-class OpenAIHelper:
+class OpenAiHelper:
     """A helper for OpenAI's API
-    
+
     It includes information and tools specifically for interacting with OpenAI's API.
     """
-
-    # see here: https://platform.openai.com/docs/models/
-    GPT_3_5 = "gpt-3.5-turbo"
-    GPT_3_5_301 = "gpt-3.5-turbo-0301"
-    GPT_3_5_16K = "gpt-3.5-turbo-16k"
-    GPT_3_5_ALL = [GPT_3_5, GPT_3_5_301, GPT_3_5_16K]
-
-    GPT_4 = "gpt-4"
-    GPT_4_32K = "gpt-4-32k"
-    GPT_4_0314 = "gpt-4-0314"
-    GPT_4_32K_0314 = "gpt-4-32k-0314"
-    GPT_4_ALL = [GPT_4, GPT_4_32K, GPT_4_0314, GPT_4_32K_0314]
-
-    GPT_DEFAULT = GPT_3_5
-
-    GPT_ALL = [*GPT_3_5_ALL, *GPT_4_ALL]
 
     GPT_3_5_MAX_TOKENS = 4_096
     GPT_4_MAX_TOKENS = 8_192
     GPT_4_32K_MAX_TOKENS = 32_768
 
-    def __init__(self, model: str, payload: List[Dict], stream=False):
-        self._model = model
-        self._payload = payload
-        self._stream = stream
+    def __init__(self, model: str, messages: Messages, stream=False):
+        self._model: str = model
+        self._messages: list[dict] = [message.to_dictionary_reduced_context() for message in messages.messages]
+        self._stream: bool = stream
 
-    def send(self) -> requests.Response:
+    def send(self) -> Union[Response | None]:
         """Sends message(s) to openai
 
         :return: The response string from openai
         """
         logger.info("Sending message to openai")
-        self._set_api_key()
-
-        key = os.environ["OPENAI_API_KEY"]
-        response = self._post_request(key=key)
+        self._export_api_key_to_environment_variable()
+        key: str = os.environ["OPENAI_API_KEY"]
+        response: Union[Response | None] = self._post_request(key=key)
 
         return response
 
-    def is_valid_api_key(self, key) -> bool:
+    def is_valid_api_key(self, key: str) -> bool:
         """Sends a POST request to the Openai API.
         We expect a return of True if we have a valid key and false if not.
         """
-        response = self._post_request(key=key)
+        response: Union[Response | None] = self._post_request(key=key)
         if response is None:
             return False
         else:
@@ -68,7 +55,7 @@ class OpenAIHelper:
 
         return is_valid
 
-    def _set_api_key(self) -> None:
+    def _export_api_key_to_environment_variable(self) -> None:
         if "OPENAI_API_KEY" not in os.environ:
             logger.info("Exporting API key to environment variable")
 
@@ -79,9 +66,8 @@ class OpenAIHelper:
         else:
             logger.info("API key already in environment variable")
 
-    def _post_request(self, key: str) -> requests.Response:
+    def _post_request(self, key: str) -> Response | None:
         logger.info("POSTing request to openai API")
-        messages = self._payload
 
         request_url = "https://api.openai.com/v1/chat/completions"
         request_headers = {
@@ -91,10 +77,10 @@ class OpenAIHelper:
         request_body = {
             "model": self._model,
             "stream": self._stream,
-            "messages": messages,
+            "messages": self._messages,
         }
 
-        response = None  # return None rather than uninitiated variable
+        response: Union[Response | None] = None  # return None rather than uninitiated variable
         try:
             response = requests.post(
                 request_url,
@@ -115,7 +101,7 @@ class OpenAIHelper:
 
         # Http error codes reference: https://platform.openai.com/docs/guides/error-codes/api-errors
         if response is not None:  # response retrieved
-            code = response.status_code
+            code: int = response.status_code
             if code >= 400:
                 match code:
                     case HTTPStatus.UNAUTHORIZED.value:  # 401