+++ gptcli/src/cli.py
@@ -3,11 +3,12 @@
 
 import argparse
 import logging
+from logging import Logger
 
 from gptcli._version import __version__
-from gptcli.src.api_helper import OpenAIHelper
+from gptcli.src.supported_models import openai
 
-logger = logging.getLogger(__name__)
+logger: Logger = logging.getLogger(__name__)
 
 
 class CommandLineInterface:
@@ -48,9 +49,9 @@ class CommandLineInterface:
         self.parser.add_argument(
             "--model",
             type=str,
-            choices=OpenAIHelper.GPT_ALL,
-            default=OpenAIHelper.GPT_DEFAULT,
-            help=f"Defaults to {OpenAIHelper.GPT_DEFAULT}. The model to use.",
+            choices=openai.values(),
+            default=openai["GPT_4O"],
+            help=f"Defaults to {openai['GPT_4O']}. The model to use.",
         )
         self.parser.add_argument(
             "--key",