+++ gptcli/src/message.py
@@ -1,13 +1,14 @@
 """Will handle messages to and from Openai's API"""
 
 import logging
-from typing import Dict, List, Set
+from logging import Logger
 
 import tiktoken
+from tiktoken import Encoding
 
-from gptcli.src.api_helper import OpenAIHelper
+from gptcli.src.supported_models import openai
 
-logger = logging.getLogger(__name__)
+logger: Logger = logging.getLogger(__name__)
 
 
 class Message:
@@ -53,11 +54,10 @@ class Message:
         """
 
         logger.info("Counting tokens for message")
-        supported_models: Set[str] = set([*(OpenAIHelper.GPT_3_5_ALL), *(OpenAIHelper.GPT_4_ALL)])
-        if self._model not in supported_models:
+        if self._model not in openai.values():
             raise NotImplementedError(f"num_tokens_from_message() is not presently implemented for {self._model}.")
         else:
-            encoding = tiktoken.encoding_for_model(self._model)
+            encoding: Encoding = self.encoding()
             num_tokens: int = 0
             num_tokens += 4  # every message follows <im_start>{role/name}\n{content}<im_end>\n
             num_tokens += len(encoding.encode(self._content))
@@ -68,6 +68,22 @@ class Message:
 
             return num_tokens
 
+    def encoding(self) -> Encoding:
+        """Determine the encoding to use for the Message.
+        The encoding is derived from the LLM model used, via tiktoken.
+        The Openai docs seem to prefer 'cl100k_base' encoding for newer models.
+        Which may be why it is used as the fallback encoding type. See the link for more:
+        https://platform.openai.com/docs/guides/text-generation/managing-tokens
+
+        Returns:
+            Encoding: The encoding of the message, which is the encoding used for the model
+        """
+        try:
+            encoding: Encoding = tiktoken.encoding_for_model(self._model)
+        except KeyError:
+            encoding = tiktoken.get_encoding("cl100k_base")
+        return encoding
+
     def to_dictionary_reduced_context(self) -> dict:
         """Use this lightweight version when sending messages to API endpoints.
         If we send less, we save tokens.
@@ -75,7 +91,6 @@ class Message:
         Returns:
             dict: A lightweight dictionary representation of the Messsage class.
         """
-
         return {
             "role": self._role,
             "content": self._content,
@@ -88,7 +103,6 @@ class Message:
         Returns:
             dict: A complete dictionary representation of the Messsage class.
         """
-
         return {
             "role": self._role,
             "content": self._content,
@@ -141,11 +155,11 @@ class Messages:
     that are not offered by Python dicitionaries or lists.
     """
 
-    def __init__(self, messages: List[Message] | None = None) -> None:
-        self._messages: List[Message] = messages if messages is not None else list()
+    def __init__(self, messages: list[Message] | None = None) -> None:
+        self._messages: list[Message] = messages if messages is not None else list()
         self._tokens = self._count_tokens()
 
-    def add_message(self, message: Message) -> None:
+    def add_message(self, message: Message | None) -> None:
         """Add a Message object to Messages.
 
         Args:
@@ -165,12 +179,12 @@ class Messages:
     def _count_tokens(self) -> int:
         logger.info("Counting total number of used in messages.")
         count: int = 0
-        for message in self.messages:
+        for message in self._messages:
             count = count + message.tokens
         return count
 
     @property
-    def messages(self) -> List[Dict]:
+    def messages(self) -> list[Message]:
         return self._messages
 
     @property